+++ lib/src/command/project.dart
@@ -80,6 +80,8 @@ class ProjectCommand extends BaseVoidCommand {
         return CMakeCompiler();
       case LibType.cMeson:
         return MesonCompiler();
+      case LibType.cMakefile:
+        return MakefileCompiler();
     }
   }
 }